+++ core-ng/src/main/java/core/framework/impl/web/HTTPHandler.java
@@ -2,7 +2,7 @@ package core.framework.impl.web;
 
 import core.framework.impl.log.ActionLog;
 import core.framework.impl.log.LogManager;
-import core.framework.impl.web.bean.BeanClassNameValidator;
+import core.framework.impl.web.bean.BeanMapperRegistry;
 import core.framework.impl.web.bean.RequestBeanMapper;
 import core.framework.impl.web.bean.ResponseBeanMapper;
 import core.framework.impl.web.controller.ControllerHolder;
@@ -40,9 +40,9 @@ public class HTTPHandler implements HttpHandler {
     public final WebContextImpl webContext = new WebContextImpl();
     public final HTTPErrorHandler errorHandler;
 
-    private final BeanClassNameValidator beanClassNameValidator = new BeanClassNameValidator();
-    public final RequestBeanMapper requestBeanMapper = new RequestBeanMapper(beanClassNameValidator);
-    public final ResponseBeanMapper responseBeanMapper = new ResponseBeanMapper(beanClassNameValidator);
+    public final BeanMapperRegistry beanMapperRegistry = new BeanMapperRegistry();
+    public final RequestBeanMapper requestBeanMapper = new RequestBeanMapper(beanMapperRegistry);
+    public final ResponseBeanMapper responseBeanMapper = new ResponseBeanMapper(beanMapperRegistry);
 
     private final Logger logger = LoggerFactory.getLogger(HTTPHandler.class);
     private final LogManager logManager;