+++ core-ng/src/main/java/core/framework/impl/web/bean/RequestBeanMapper.java
@@ -1,6 +1,5 @@
 package core.framework.impl.web.bean;
 
-import core.framework.api.json.Property;
 import core.framework.api.web.service.QueryParam;
 import core.framework.impl.validate.Validator;
 import core.framework.util.Maps;
@@ -11,12 +10,11 @@ import java.util.Map;
  * @author neo
  */
 public class RequestBeanMapper {
-    private final Map<Class<?>, BeanMapper<?>> requestBeanMappers = Maps.newConcurrentHashMap();
+    private final BeanMapperRegistry registry;
     private final Map<Class<?>, QueryParamMapperHolder<?>> queryParamMappers = Maps.newConcurrentHashMap();
-    private final BeanClassNameValidator classNameValidator;
 
-    public RequestBeanMapper(BeanClassNameValidator classNameValidator) {
-        this.classNameValidator = classNameValidator;
+    public RequestBeanMapper(BeanMapperRegistry registry) {
+        this.registry = registry;
     }
 
     public <T> Map<String, String> toParams(Class<T> beanClass, T bean) {
@@ -35,7 +33,7 @@ public class RequestBeanMapper {
     @SuppressWarnings("unchecked")
     public <T> QueryParamMapperHolder<T> registerQueryParamBean(Class<T> beanClass) {
         return (QueryParamMapperHolder<T>) queryParamMappers.computeIfAbsent(beanClass, key -> {
-            new QueryParamBeanClassValidator(beanClass, classNameValidator).validate();
+            new QueryParamBeanClassValidator(beanClass, registry).validate();
             QueryParamMapper<T> mapper = new QueryParamMapperBuilder<>(beanClass).build();
             var validator = new Validator(beanClass, field -> field.getDeclaredAnnotation(QueryParam.class).name());
             return new QueryParamMapperHolder<>(mapper, validator);
@@ -43,24 +41,20 @@ public class RequestBeanMapper {
     }
 
     public <T> byte[] toJSON(Class<T> beanClass, T bean) {
-        BeanMapper<T> holder = registerRequestBean(beanClass);
-        holder.validator.validate(bean, false);
-        return holder.writer.toJSON(bean);
+        BeanMapper<T> mapper = registerRequestBean(beanClass);
+        mapper.validator.validate(bean, false);
+        return mapper.writer.toJSON(bean);
     }
 
     public <T> T fromJSON(Class<T> beanClass, byte[] body) {
-        BeanMapper<T> holder = registerRequestBean(beanClass);
-        T bean = holder.reader.fromJSON(body);
-        holder.validator.validate(bean, false);
+        BeanMapper<T> mapper = registerRequestBean(beanClass);
+        T bean = mapper.reader.fromJSON(body);
+        mapper.validator.validate(bean, false);
         return bean;
     }
 
-    @SuppressWarnings("unchecked")
     public <T> BeanMapper<T> registerRequestBean(Class<T> beanClass) {
-        return (BeanMapper<T>) requestBeanMappers.computeIfAbsent(beanClass, type -> {
-            new BeanClassValidator(beanClass, classNameValidator).validate();
-            return new BeanMapper<>(beanClass, new Validator(beanClass, field -> field.getDeclaredAnnotation(Property.class).name()));
-        });
+        return registry.register(beanClass);
     }
 
     static class QueryParamMapperHolder<T> {