+++ core-ng/src/main/java/core/framework/impl/web/management/APIController.java
@@ -8,16 +8,19 @@ import core.framework.web.Request;
 import core.framework.web.Response;
 
 import java.util.Map;
+import java.util.Set;
 
 /**
  * @author neo
  */
 public class APIController implements Controller {
     private final Map<String, Class<?>> serviceInterfaces;
+    private final Set<Class<?>> beanClasses;
     private final IPAccessControl accessControl;
 
-    public APIController(Map<String, Class<?>> serviceInterfaces, IPAccessControl accessControl) {
+    public APIController(Map<String, Class<?>> serviceInterfaces, Set<Class<?>> beanClasses, IPAccessControl accessControl) {
         this.serviceInterfaces = serviceInterfaces;
+        this.beanClasses = beanClasses;
         this.accessControl = accessControl;
     }
 
@@ -27,6 +30,7 @@ public class APIController implements Controller {
 
         var builder = new APIDefinitionBuilder();
         serviceInterfaces.values().forEach(builder::addServiceInterface);
+        beanClasses.forEach(builder::parseBeanType);
         APIDefinitionResponse response = builder.build();
 
         return Response.bean(response);