+++ core-ng/src/test/java/core/framework/impl/web/response/BeanBodyTest.java
@@ -1,7 +1,7 @@
 package core.framework.impl.web.response;
 
 import core.framework.impl.validate.ValidationException;
-import core.framework.impl.web.bean.BeanClassNameValidator;
+import core.framework.impl.web.bean.BeanMapperRegistry;
 import core.framework.impl.web.bean.ResponseBeanMapper;
 import core.framework.impl.web.bean.TestBean;
 import io.undertow.io.Sender;
@@ -17,7 +17,7 @@ class BeanBodyTest {
     @Test
     void send() {
         var sender = mock(Sender.class);
-        var context = new ResponseHandlerContext(new ResponseBeanMapper(new BeanClassNameValidator()), null);
+        var context = new ResponseHandlerContext(new ResponseBeanMapper(new BeanMapperRegistry()), null);
         var body = new BeanBody(new TestBean());
         assertThatThrownBy(() -> body.send(sender, context))
                 .isInstanceOf(ValidationException.class);